Organized component file structure
